Finnished scraping. Backed up Db in BackUpDataBase, old back up should be updated to the new one. Deleted SmartPotV4 for remake.

+++ SQLserverFinale/SQLserverFinale/Models/Plant.cs
@@ -29,8 +29,8 @@ namespace SQLserverFinale.Models
         [MaxLength(100)]
         public string FlowerColor { get; set; }
 
-        public int PhMinVal { get; set; }
-        public int PhMaxVal { get; set; }
+        public float PhMinVal { get; set; }
+        public float PhMaxVal { get; set; }
         public int MinTemp { get; set; }
         public int MaxTemp { get; set; }
         public int SunReq { get; set; }
